reuse Violations page skill URL parser for agentic assets
Extract deriveSkillOrToolName / skillNameFromUrl into a shared helper so skill attribution does not duplicate the existing violations-page logic.
